src/slurp.tcl:  * List of header fields is now reformatted and handed off to the message    body processing command.  * Message body is unmarshaled.
NOTE: it turns out that in fact we have to implement some kind of "message objects" --
variables that hold the message state and are handed off to code that actually
wants them to process.
Since, strictly speaking, an unmarshaled message can be anything, it's wise to
just hand it to some upper-level code to decide what to do next.